+++ core/amber/src/main/scala/edu/uci/ics/amber/engine/architecture/controller/execution/WorkerPortExecution.scala
@@ -1,4 +1,4 @@
-package edu.uci.ics.amber.engine.architecture.controller
+package edu.uci.ics.amber.engine.architecture.controller.execution
 
 class WorkerPortExecution extends Serializable {
   var completed: Boolean = false
